section6/app/requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==1.4.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
